# frontend/landings/agivys/landing/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgiVysSystem API",
    "description": "API para gestão do ecossistema AgiVysSystem",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://joederblanca.com.br/agivys-api"
    }
  ],
  "paths": {
...
    "/api/v1/checks/check-email/{email}": {
      "get": {
        "tags": [
          "Checks"
        ],
        "summary": "Verificação de Disponibilidade de E-mail",
        "description": "Consulta na base de dados (Identity) se o e-mail fornecido já está vinculado a alguma conta de usuário existente.\r\nEste endpoint é público e frequentemente utilizado durante etapas de pré-cadastro (Onboarding) para validação assíncrona de formulários em tempo real.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "O endereço de e-mail completo a ser verificado.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requisição bem-sucedida. Retorna um objeto indicando se o e-mail já existe (`true` ou `false`)."
          },
          "400": {
            "description": "Formato de e-mail inválido ou parâmetro ausente."
          },
          "500": {
            "description": "Erro interno do servidor ao tentar consultar o banco de dados."
          }
        }
      }
    },
    "/api/v1/checks/check-cpf/{document}": {
      "get": {
        "tags": [
          "Checks"
        ],
        "summary": "Verificação de Disponibilidade de CPF",
        "description": "Verifica se o documento (CPF) informado já possui um cadastro atrelado na tabela de Pessoas (Person).\r\nEndpoint público de utilidade para formulários. Suporta o recebimento do documento com ou sem máscara de formatação (pontos e traços), uma vez que a API higieniza a string mantendo apenas números.",
        "parameters": [
          {
            "name": "document",
            "in": "path",
            "description": "O número do CPF a ser consultado.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requisição bem-sucedida. Retorna um objeto indicando se o documento já existe (`true` ou `false`)."
          },
          "400": {
            "description": "CPF com tamanho inválido ou string malformada."
          },
          "500": {
            "description": "Erro interno do servidor ao tentar consultar o banco de dados."
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Cole APENAS o token JWT.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}